+++ src/mocks/RuleEngineExposed.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MPL-2.0
 pragma solidity ^0.8.20;
 
-import {RuleEngine} from "../RuleEngine.sol";
-import {RuleEngineOwnable} from "../RuleEngineOwnable.sol";
+import {RuleEngine} from "../deployment/RuleEngine.sol";
+import {RuleEngineOwnable} from "../deployment/RuleEngineOwnable.sol";
 
 /**
  * @title RuleEngineExposed